+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This update accompanies the release of the [Telegram Bot API 6.6](https://core.t
    see [setMyShortDescription](https://core.telegram.org/bots/api#setmyshortdescription) for more information.
  * Added the ability to get the current bot short description in the given language as the class [BotShortDescription](https://core.telegram.org/bots/api#botshortdescription)
    using the method `\TgBotLib\Bot > getMyShortDescription()` see [getMyShortDescription](https://core.telegram.org/bots/api#getmyshortdescription) for more information.
+ * Added method `\TgBotLib\Bot > sendSticker()` to send a sticker to a chat. See [sendSticker](https://core.telegram.org/bots/api#sendsticker) for more information.
 
 ### Changed
  * Removed unused `__destruct()` method from `\TgBotLib\Bot`